Move the maven-gpg-plugin to own profile
Move the maven-gpg-plugin execution (which is always being executed on
mvn install) to an own profile so that external contributors can disable
it. This is necessary for testing locally (with mvn install and then
building a test project).

Plugin is still activated by default to avoid breaking any processes I
cannot see from the contributor position.