Make two tests also working on windows
I wanted to add a windows ci pipeline. At the moment not possible
because some tests are killing an Axon Ivy Engine with SIGKILL. This
prevents the Axon Ivy Engine from stoping the bundled Elasticsearch
Service, which than makes it impossible to clean the engine directory.

Seems that it is not possible with commons.cli to gracefully stop
a service. I will need first to use the new Java Process API to manage
the Axon Ivy Engine Process. This will also eliminate some maven warnings
because the process is killed with SIGKILL.